third_party/blink/web_tests/http/tests/devtools/startup/sources/debugger/script-formatter-breakpoints-inline-with-sourceURL-expected.txt

Tests the script formatting is working with breakpoints for inline scripts with #sourceURL=.


Running: testSetup

Running: testBreakpointsInOriginalAndFormattedSource
Script execution paused.
Breakpoint sidebar pane while paused in raw
named-inline-script.js:5checked breakpoint hit      console.log("Hello!");
Script execution resumed.
Script execution paused.
Breakpoint sidebar pane while paused in pretty printed
named-inline-script.js:formatted:3checked breakpoint hit    console.log("Hello!");
Breakpoint sidebar pane while paused after removing breakpoint in pretty printed and closing pretty printed
No breakpoints
Breakpoint sidebar pane while paused in original script again
named-inline-script.js:5checked breakpoint hit      console.log("Hello!");
Script execution resumed.
